\name{jumpstart}
\alias{jumpstart}
\title{
  Get the starting values of the free parameters of a model.
}
\description{
  Get the starting values of the free parameters of a model.
}
\usage{
jumpstart(model)
}
\arguments{
  \item{model}{
    A lavaanified model.
}
}
\value{
  A vector of free parameters.
}
