\name{mirror.tri}
\alias{mirror.tri}
\title{
  Place the values below or above the diagonal of a square matrix in
  the opposite triangle.
}
\description{
  This function gets the values below or above the diagonal of a
  square matrix and transposes them.
}
\usage{
mirror.tri(mat, upper.on.lower = FALSE)
}
\arguments{
  \item{mat}{
    A square matrix.
  }
  \item{upper.on.lower}{
    Should the upper values be transposed to the triangle below the
    diagonal?
  }
}
